[refactorying] detail
